Deleted the html & worked on Code Coverage
cleaned the html out of the src folder. and wrote a bunch of unit tests.

+++ ArgumentParser/src/main/java/edu/jsu/mcis/ArgumentParser.java
@@ -505,7 +505,7 @@ public class ArgumentParser {
 						}
 						errorValue = Integer.toString(intValue);
 					}
-					else if(currentArg.getType().equals("integer")){
+					else if(currentArg.getType().equals("float")){
 						float floatValue = (Float)currentArg.getValue();
 						if(valueSet[j].equals(Float.toString(floatValue))){
 							valueSetContainsArgValue = true;
